GenerateData/requirements.txt
pyqt5
pillow
flask
websockets